backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
psycopg2>=2.9.9
redis==5.0.1
celery==5.3.4
Pillow>=10.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
stripe==7.4.0
razorpay==1.4.1
boto3==1.34.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.3.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.0.0
python-multipart==0.0.6
requests==2.31.0
cryptography==41.0.7 